requirements.txt
flask
langchain
langchain-anthropic
pillow
pytesseract
python-dotenv
flask-debugtoolbar
flask-session